Module for polling functionality
